+++ UnityProject/Assets/Scripts/Messages/Server/FinishedAddedObserverMessage.cs.meta
